Renaming files to follow a more modern TS convention
