+++ server/.vscode/settings.json
@@ -24,5 +24,6 @@
     "titleBar.activeForeground": "#e7e7e7",
     "titleBar.inactiveBackground": "#ff000099",
     "titleBar.inactiveForeground": "#e7e7e799"
-  }
+  },
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
