+++ server/src/__tests__/integration/role.integration.test.ts
@@ -2,7 +2,7 @@ import request from 'supertest';
 import { afterAll, beforeEach, describe, expect, it } from 'vitest';
 import app from '../../app';
 import prisma from '../../prisma';
-import { generateJwtToken } from '../../utils/generate-token';
+import { generateJwtToken } from '../../utils/generate-token.util';
 
 const authCookie = `token=${generateJwtToken(1, 'Admin')}`;
 const roleApiRoute = '/api/v1/roles';
