+++ server/src/__tests__/role-service.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { RoleService } from '../services/role-service';
-import { NotFoundError } from '../errors/not-found-error';
+import { RoleService } from '../services/role.service';
+import { NotFoundError } from '../errors/not-found.error';
 import { Role } from '../types/role';
 
 const mockPrisma = {
