+++ server/src/errors/bad-request.error.ts
@@ -1,4 +1,4 @@
-import { AppError } from './app-error';
+import { AppError } from './app.error';
 
 export class BadRequest extends AppError {
   constructor(message: string) {
