+++ server/src/errors/conflict.error.ts
@@ -1,4 +1,4 @@
-import { AppError } from './app-error';
+import { AppError } from './app.error';
 
 export class ConflictError extends AppError {
   constructor(message: string) {
