+++ server/src/errors/not-found.error.ts
@@ -1,4 +1,4 @@
-import { AppError } from './app-error';
+import { AppError } from './app.error';
 
 export class NotFoundError extends AppError {
   constructor(message: string) {
