+++ server/src/errors/unauthorized.error.ts
@@ -1,4 +1,4 @@
-import { AppError } from './app-error';
+import { AppError } from './app.error';
 
 export class UnauthorizedError extends AppError {
   constructor(message: string) {
