+++ server/src/middleware/auth-token.middleware.ts
