+++ server/src/middleware/error-handler.middleware.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
-import logger from '../utils/logger';
-import { AppError } from '../errors/app-error';
+import logger from '../utils/logger.util';
+import { AppError } from '../errors/app.error';
 
 export const errorHandler = (
   error: unknown,
