+++ server/src/routes/role.route.ts
@@ -1,12 +1,12 @@
-import express from 'express';
-import { RoleService } from '../services/role-service';
-import { RoleController } from '../controllers/role-controller';
-import { asyncWrapper } from '../utils/async-wrapper';
+import express, { Router } from 'express';
+import { RoleService } from '../services/role.service';
+import { RoleController } from '../controllers/role.controller';
+import { asyncWrapper } from '../utils/async-wrapper.util';
 import { PrismaClient } from '@prisma/client';
-import { authenticateToken } from '../middleware/auth-token';
-import { checkRole } from '../middleware/check-role';
+import { authenticateToken } from '../middleware/auth-token.middleware';
+import { checkRole } from '../middleware/check-role.middleware';
 
-export const createRoleRouter = (prisma: PrismaClient) => {
+export const createRoleRouter = (prisma: PrismaClient): Router => {
   const router = express.Router();
   const roleService = new RoleService(prisma);
   const roleController = new RoleController(roleService);
