+++ server/src/routes/take.route.ts
@@ -1,11 +1,11 @@
-import express from 'express';
-import { TakeService } from '../services/take-service';
-import { TakeController } from '../controllers/take-controller';
-import { asyncWrapper } from '../utils/async-wrapper';
+import express, { Router } from 'express';
+import { TakeService } from '../services/take.service';
+import { TakeController } from '../controllers/take.controller';
+import { asyncWrapper } from '../utils/async-wrapper.util';
 import { PrismaClient } from '@prisma/client';
-import { authenticateToken } from '../middleware/auth-token';
+import { authenticateToken } from '../middleware/auth-token.middleware';
 
-export const createTakeRouter = (prisma: PrismaClient) => {
+export const createTakeRouter = (prisma: PrismaClient): Router => {
   const router = express.Router();
   const takeService = new TakeService(prisma);
   const takeController = new TakeController(takeService);
