+++ server/src/routes/user.route.ts
@@ -1,12 +1,12 @@
-import express from 'express';
-import { UserService } from '../services/user-service';
-import { UserController } from '../controllers/user-controller';
-import { asyncWrapper } from '../utils/async-wrapper';
-import { authenticateToken } from '../middleware/auth-token';
-import { checkRole } from '../middleware/check-role';
+import express, { Router } from 'express';
+import { UserService } from '../services/user.service';
+import { UserController } from '../controllers/user.controller';
+import { asyncWrapper } from '../utils/async-wrapper.util';
+import { authenticateToken } from '../middleware/auth-token.middleware';
+import { checkRole } from '../middleware/check-role.middleware';
 import { PrismaClient } from '@prisma/client';
 
-export const createUserRouter = (prisma: PrismaClient) => {
+export const createUserRouter = (prisma: PrismaClient): Router => {
   const router = express.Router();
   router.use(authenticateToken, checkRole('admin'));
