+++ server/src/services/email.service.ts
@@ -1,7 +1,14 @@
-import nodemailer from 'nodemailer';
+import nodemailer, { SentMessageInfo } from 'nodemailer';
 import fs from 'fs';
 import path from 'path';
 import Handlebars from 'handlebars';
+
+interface EmailOptions {
+  to: string;
+  subject: string;
+  text?: string;
+  html?: string;
+}
 
 export const transporter = nodemailer.createTransport({
   host: process.env.SMTP_HOST,
@@ -13,18 +20,11 @@ export const transporter = nodemailer.createTransport({
   },
 });
 
-export async function sendEmail({
-  to,
-  subject,
-  text,
-  html,
-}: {
-  to: string;
-  subject: string;
-  text?: string;
-  html?: string;
-}) {
+export const sendEmail = async (
+  options: EmailOptions,
+): Promise<SentMessageInfo> => {
   try {
+    const { to, subject, text, html } = options;
     const info = await transporter.sendMail({
       from: process.env.SMTP_USER, // Should be change later
       to,
@@ -38,24 +38,24 @@ export async function sendEmail({
     console.error('Failed to send email:', error);
     throw error;
   }
-}
+};
 
-export function renderTemplate(
+export const renderTemplate = <T extends Record<string, unknown>>(
   templateName: string,
-  data: Record<string, any>,
-): string {
+  data: T,
+): string => {
   const templatePath = path.join(__dirname, '../templates', templateName);
   const templateSource = fs.readFileSync(templatePath, 'utf8');
   const template = Handlebars.compile(templateSource);
   return template(data);
-}
+};
 
-export async function sendCodeEmail(
+export const sendCodeEmail = async (
   to: string,
   nickname: string,
   code: string,
-) {
+): Promise<SentMessageInfo> => {
   const html = renderTemplate('code-email.html', { nickname, code });
 
   return sendEmail({ to, subject: 'Welcome to Hot Takes!', html });
-}
+};
