+++ server/src/utils/format-user.util.ts
