+++ server/src/utils/generate-code.util.ts
