+++ server/src/utils/generate-token.util.ts
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
 
-export const generateJwtToken = (userId: number, role: string) => {
+export const generateJwtToken = (userId: number, role: string): string => {
   return jwt.sign(
     { userId, role },
     process.env.JWT_SECRET as string,
